moved processor to dom preprocessor
