+++ src/Converter/ConfluenceConverter.php
@@ -20,7 +20,8 @@ use HalloWelt\MigrateConfluence\Converter\Postprocessor\NestedHeadings;
 use HalloWelt\MigrateConfluence\Converter\Postprocessor\RestorePStyleTag;
 use HalloWelt\MigrateConfluence\Converter\Postprocessor\RestoreTimeTag;
 use HalloWelt\MigrateConfluence\Converter\Postprocessor\TasksReportMacro as RestoreTasksReportMacro;
-use HalloWelt\MigrateConfluence\Converter\Preprocessor\CDATAClosingFixer;
+use HalloWelt\MigrateConfluence\Converter\Preprocessor\dom\SanitizeLinkContent;
+use HalloWelt\MigrateConfluence\Converter\Preprocessor\html\CDATAClosingFixer;
 use HalloWelt\MigrateConfluence\Converter\Processor\AlignMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\AnchorLink;
 use HalloWelt\MigrateConfluence\Converter\Processor\AnchorMacro;
@@ -37,7 +38,6 @@ use HalloWelt\MigrateConfluence\Converter\Processor\Emoticon;
 use HalloWelt\MigrateConfluence\Converter\Processor\ExcerptIncludeMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\ExcerptMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\ExpandMacro;
-use HalloWelt\MigrateConfluence\Converter\Processor\ExtractComplexLinkContent;
 use HalloWelt\MigrateConfluence\Converter\Processor\GalleryMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\GliffyMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\Image;
@@ -333,7 +333,6 @@ class ConfluenceConverter extends PandocHTML implements IOutputAwareInterface {
 			new ExcerptIncludeMacro( $this->dataLookup, $this->currentSpace ),
 			new Emoticon(),
 			new PreserveTasksReportMacro( $this->dataLookup ),
-			new ExtractComplexLinkContent(),
 			new Image(
 				$this->dataLookup, $this->currentSpace, $currentPageTitle, $this->advancedConfig
 			),
@@ -511,6 +510,8 @@ class ConfluenceConverter extends PandocHTML implements IOutputAwareInterface {
 			throw new Exception( 'Unconvertable' );
 		}
 
+		$this->preprocessDomSource( $dom );
+
 		$preprocessedPathname = str_replace( '.mraw', '.mprep', $this->rawFile->getPathname() );
 		$dom->saveHTMLFile( $preprocessedPathname );
 		$this->preprocessedFile = new SplFileInfo( $preprocessedPathname );
@@ -530,7 +531,7 @@ class ConfluenceConverter extends PandocHTML implements IOutputAwareInterface {
 		$preprocessors = [
 			new CDATAClosingFixer()
 		];
-		/** @var IPreprocessor $preprocessor */
+		/** @var IHtmlPreprocessor $preprocessor */
 		foreach ( $preprocessors as $preprocessor ) {
 			$sContent = $preprocessor->preprocess( $sContent );
 		}
@@ -568,6 +569,21 @@ class ConfluenceConverter extends PandocHTML implements IOutputAwareInterface {
 		$sContent = '<xml xmlns:ac="some" xmlns:ri="thing" xmlns:bs="bluespice">' . $sContent . '</xml>';
 
 		return $sContent;
+	}
+
+	/**
+	 * @param DOMDocument $dom
+	 * @return void
+	 */
+	protected function preprocessDomSource( DOMDocument $dom ): void {
+		$preprocessors = [
+			new SanitizeLinkContent()
+		];
+
+		/** @var IHtmlPreprocessor $preprocessor */
+		foreach ( $preprocessors as $preprocessor ) {
+			$preprocessor->preprocess( $dom );
+		}
 	}
 
 	/**
