+++ src/Converter/IDomPreprocessor.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace HalloWelt\MigrateConfluence\Converter;
+
+use DOMDocument;
+
+interface IDomPreprocessor {
+
+	/**
+	 * @param DOMDocument $dom
+	 *
+	 * @return void
+	 */
+	public function preprocess( DOMDocument $dom ): void;
+}
